# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Resource-Service is service for importing kubernetes resources to other services.",
    "title": "Kube-Importer",
    "version": "0.0.1"
  },
  "paths": {
    "/all": {
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Import all resources.",
        "operationId": "ImportAllHandler",
        "responses": {
          "202": {
            "description": "import result",
            "schema": {
              "$ref": "#/definitions/ImportResponseTotal"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/all/ws": {
      "get": {
        "tags": [
          "Import"
        ],
        "summary": "Import all resources with websockets response.",
        "operationId": "ImportAllWS",
        "parameters": [
          {
            "type": "string",
            "format": "string",
            "name": "Upgrade",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "string",
            "name": "Connection",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "string",
            "name": "Sec-WebSocket-Key",
            "in": "header",
            "required": true
          },
          {
            "type": "integer",
            "format": "int",
            "name": "Sec-Websocket-Version",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "101": {
            "description": "import response"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/configmaps": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export config maps.",
        "operationId": "ExportConfigMapsList",
        "responses": {
          "200": {
            "description": "export result",
            "schema": {
              "$ref": "#/definitions/ConfigMapsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Import configmaps.",
        "operationId": "ImportConfigMapsList",
        "responses": {
          "202": {
            "description": "import result",
            "schema": {
              "$ref": "#/definitions/ImportResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/deployments": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export namespaces.",
        "operationId": "ExportDeploymentsList",
        "responses": {
          "200": {
            "description": "export result",
            "schema": {
              "$ref": "#/definitions/DeploymentsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Import deployments.",
        "operationId": "ImportDeploymentsList",
        "responses": {
          "202": {
            "description": "import result",
            "schema": {
              "$ref": "#/definitions/ImportResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/ingresses": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export ingresses.",
        "operationId": "ExportIngressesList",
        "responses": {
          "200": {
            "description": "export result",
            "schema": {
              "$ref": "#/definitions/ServiceWithParamList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Import ingresses.",
        "operationId": "ImportIngressesList",
        "responses": {
          "202": {
            "description": "import result",
            "schema": {
              "$ref": "#/definitions/ImportResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/namespaces": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export namespaces.",
        "operationId": "ExportNamespacesList",
        "responses": {
          "200": {
            "description": "export result",
            "schema": {
              "$ref": "#/definitions/NamespacesList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Import namespaces.",
        "operationId": "ImportNamespacesList",
        "responses": {
          "202": {
            "description": "import result",
            "schema": {
              "$ref": "#/definitions/ImportResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/services": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export services.",
        "operationId": "ExportServicesList",
        "responses": {
          "200": {
            "description": "export result",
            "schema": {
              "$ref": "#/definitions/ServiceWithParamList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Import services.",
        "operationId": "ImportServicesList",
        "responses": {
          "202": {
            "description": "import result",
            "schema": {
              "$ref": "#/definitions/ImportResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/storages": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export storages.",
        "operationId": "ExportStoragesList",
        "responses": {
          "200": {
            "description": "export result",
            "schema": {
              "$ref": "#/definitions/StorageList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Import storages.",
        "operationId": "ImportStoragesList",
        "responses": {
          "202": {
            "description": "import result",
            "schema": {
              "$ref": "#/definitions/ImportResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/volumes": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export volumes.",
        "operationId": "ExportVolumesList",
        "responses": {
          "200": {
            "description": "export result",
            "schema": {
              "$ref": "#/definitions/VolumesList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Import volumes.",
        "operationId": "ImportVolumesList",
        "responses": {
          "202": {
            "description": "import result",
            "schema": {
              "$ref": "#/definitions/ImportResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/importAllHandlerDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    }
  },
  "definitions": {
...
    "ConfigMap": {
      "description": "ConfigMap -- model for config map",
      "type": "object",
      "required": [
        "name",
        "data"
      ],
      "properties": {
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "data": {
          "$ref": "#/definitions/ConfigMapData"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ConfigMapData": {
      "description": "ConfigMapData -- model for config map data",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ConfigMapsList": {
      "description": "ConfigMapsList -- model for config maps list",
      "type": "object",
      "properties": {
        "configmaps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConfigMap"
          },
          "x-go-name": "ConfigMaps"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "Deployment": {
      "description": "Deployment -- model for deployments",
      "type": "object",
      "required": [
        "containers",
        "name",
        "replicas"
      ],
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "containers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Container"
          },
          "x-go-name": "Containers"
        },
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "image_pull_secret": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "ImagePullSecrets"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Replicas"
        },
        "solution_id": {
          "description": "Solution ID (only if deployment is part of solution)",
          "type": "string",
          "x-go-name": "SolutionID"
        },
        "status": {
          "$ref": "#/definitions/DeploymentStatus"
        },
        "total_cpu": {
          "description": "total CPU usage by all containers in this deployment",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "TotalCPU"
        },
        "total_memory": {
          "description": "total RAM usage by all containers in this deployment",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "TotalMemory"
        },
        "version": {
          "$ref": "#/definitions/Version"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "DeploymentStatus": {
      "description": "DeploymentStatus -- kubernetes status of deployment",
      "type": "object",
      "properties": {
        "available_replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AvailableReplicas"
        },
        "ready_replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ReadyReplicas"
        },
        "replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Replicas"
        },
        "unavailable_replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UnavailableReplicas"
        },
        "updated_replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdatedReplicas"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "DeploymentsList": {
      "description": "DeploymentsList -- model for deployments list",
      "type": "object",
      "properties": {
        "deployments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Deployment"
          },
          "x-go-name": "Deployments"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ImportResponse": {
      "description": "ImportResponse -- response after resources import",
      "type": "object",
      "properties": {
        "failed": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImportResult"
          },
          "x-go-name": "Failed"
        },
        "imported": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImportResult"
          },
          "x-go-name": "Imported"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ImportResponseTotal": {
      "description": "ImportResponseTotal -- import response for all services",
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/ImportResponse"
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "ImportResult": {
      "description": "ImportResult -- import result for one resource",
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "Namespace": {
      "description": "Namespace -- namespace representation",
      "type": "object",
      "required": [
        "resources"
      ],
      "properties": {
        "access": {
          "$ref": "#/definitions/AccessLevel"
        },
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "label": {
          "description": "user-visible label for the namespace",
          "type": "string",
          "x-go-name": "Label"
        },
        "max_ext_service": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "MaxExtService"
        },
        "max_int_service": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "MaxIntService"
        },
        "max_traffic": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "MaxTraffic"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "owner_login": {
          "type": "string",
          "x-go-name": "OwnerLogin"
        },
        "resources": {
          "$ref": "#/definitions/Resources"
        },
        "tariff_id": {
          "type": "string",
          "x-go-name": "TariffID"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserAccess"
          },
          "x-go-name": "Users"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "NamespacesList": {
      "description": "NamespacesList -- model for namespaces list",
      "type": "object",
      "properties": {
        "namespaces": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Namespace"
          },
          "x-go-name": "Namespaces"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ServiceWithParam": {
      "description": "ServiceWithParam -- model for service with owner",
      "type": "object",
      "required": [
        "name",
        "ports"
      ],
      "properties": {
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "deploy": {
          "type": "string",
          "x-go-name": "Deploy"
        },
        "domain": {
          "type": "string",
          "x-go-name": "Domain"
        },
        "hidden": {
          "description": "hide service from users",
          "type": "boolean",
          "x-go-name": "Hidden"
        },
        "ips": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "IPs"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "ports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServicePort"
          },
          "x-go-name": "Ports"
        },
        "solution_id": {
          "description": "Solution ID (only if service is part of solution)",
          "type": "string",
          "x-go-name": "SolutionID"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/git.containerum.net/ch/kube-api/pkg/model"
    },
    "ServiceWithParamList": {
      "description": "ServiceWithParamList -- model for services list",
      "type": "object",
      "properties": {
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServiceWithParam"
          },
          "x-go-name": "Services"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/git.containerum.net/ch/kube-api/pkg/model"
    },
...
    "StorageList": {
      "description": "StorageList -- model for storage names list",
      "type": "array",
      "items": {
        "type": "string"
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/git.containerum.net/ch/kube-api/pkg/model"
    },
...
    "VolumesList": {
      "description": "VolumesList -- model for volumes list",
      "type": "object",
      "properties": {
        "volumes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Volume"
          },
          "x-go-name": "Volumes"
        }
      },
      "x-go-package": "github.com/containerum/kube-importer/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "importAllHandlerDefaultBody": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "description": "Fields -- represents additional key-value fields of error",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "id": {
          "description": "ErrID -- represents unique error ID",
          "type": "object",
          "properties": {
            "kind": {
              "description": "ErrKind -- represents kind of error",
              "type": "integer",
              "format": "uint64",
              "x-go-package": "github.com/containerum/cherry"
            },
            "sid": {
              "description": "ErrSID -- represents service ID of error",
              "type": "string",
              "x-go-package": "github.com/containerum/cherry"
            }
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-gen-location": "operations",
      "x-go-package": "github.com/containerum/cherry"
    }
  },
...
}